feat: update users management function for admin
