+++ package.json
@@ -21,7 +21,7 @@
     "@mui/lab": "^5.0.0-alpha.88",
     "@mui/material": "^5.8.6",
     "@mui/styled-engine-sc": "^5.8.0",
-    "@mui/x-data-grid": "^5.12.3",
+    "@mui/x-data-grid": "^5.13.1",
     "@mui/x-date-pickers": "^5.0.0-beta.0",
     "@react-oauth/google": "^0.2.6",
     "@reduxjs/toolkit": "^1.8.2",
