--- pages/dashboard/admin-features/index.js
@@ -1,7 +1,0 @@
-import React from 'react'
-import AdminFeatures from 'src/components/dashboard/pages/admin-features'
-import Page from 'src/components/dashboard/main'
-
-export default function DashBoard() {
-  return <Page CurrentComponent={AdminFeatures}/>;
-}
